requirements.txt
fastapi==0.115.0
gunicorn==23.0.0
uvicorn==0.30.6
google-generativeai==0.3.0
python-dotenv==1.0.0
requests==2.31.0
scrapy==2.11.2
